+++ message_ix/models.py
@@ -3,9 +3,11 @@ from collections import ChainMap
 from copy import copy, deepcopy
 from functools import lru_cache
 from pathlib import Path
+from warnings import warn
 
 import ixmp.model.gams
 from ixmp import config
+from ixmp.utils import maybe_check_out, maybe_commit
 
 from .macro import MACRO_ITEMS
 
@@ -27,6 +29,7 @@ _ABBREV = {
     "g": ("grade", "grade"),
     "l": ("level", "level"),
     "m": ("mode", "mode"),
+    "ms": ("mode", "storage_mode"),
     "n": ("node", "node"),
     "nd": ("node", "node_dest"),
     "nl": ("node", "node_loc"),
@@ -36,6 +39,7 @@ _ABBREV = {
     "r": ("rating", "rating"),
     "s": ("shares", "shares"),
     "t": ("technology", "technology"),
+    "ts": ("technology", "storage_tec"),
     "h": ("time", "time"),
     "hd": ("time", "time_dest"),
     "ho": ("time", "time_origin"),
@@ -134,7 +138,7 @@ MESSAGE_ITEMS = {
     ),
     "map_tec_addon": dict(ix_type="set", idx_sets=["technology", "type_addon"]),
     # Mapping of storage reservoir to charger/discharger
-    "map_tec_storage": item("set", "n t storage_tec l c"),
+    "map_tec_storage": item("set", "n t m ts ms l c lvl_temporal"),
     "map_temporal_hierarchy": dict(
         ix_type="set",
         idx_sets=["lvl_temporal", "time", "time"],
@@ -245,23 +249,23 @@ MESSAGE_ITEMS = {
     "soft_new_capacity_lo": item("par", "nl t yv"),
     "soft_new_capacity_up": item("par", "nl t yv"),
     # Initial amount of storage
-    "storage_initial": item("par", "n t l c y h"),
+    "storage_initial": item("par", "n t m l c y h"),
     # Storage losses as a percentage of installed capacity
-    "storage_self_discharge": item("par", "n t l c y h"),
+    "storage_self_discharge": item("par", "n t m l c y h"),
     "subsidy": item("par", "nl type_tec ya"),
     "tax_emission": dict(
         ix_type="par", idx_sets=["node", "type_emission", "type_tec", "type_year"]
     ),
     "tax": item("par", "nl type_tec ya"),
     "technical_lifetime": item("par", "nl t yv"),
-    # Order of sub-annual time steps
+    # Order of sub-annual time slices
     "time_order": dict(ix_type="par", idx_sets=["lvl_temporal", "time"]),
     "var_cost": item("par", "nl t yv ya m h"),
     #
-    # commented: for both variables and equations, ixmp_source requires that the
+    # commented: for certain variables and equations, ixmp_source requires that the
     # `idx_sets` and `idx_names` parameters be empty, but then internally uses the
     # correct sets and names to initialize.
-    # TODO adjust ixmp_source to accept these values, then uncomment.
+    # TODO adjust JDBCBackend to accept these values (or replace it), then uncomment.
     #
     # Variables
     #
@@ -283,6 +287,9 @@ MESSAGE_ITEMS = {
     # "REL": item("var", "relation nr yr"),
     # # Stock
     # "STOCK": item("var", "n c l y"),
+    #
+    "STORAGE": item("var", "n t m l c y h"),
+    "STORAGE_CHARGE": item("var", "n t m l c y h"),
     #
     # # Equations
     # # Commodity balance
@@ -320,23 +327,6 @@ class GAMSModel(ixmp.model.gams.GAMSModel):
         },
         ixmp.model.gams.GAMSModel.defaults,
     )
-
-    @classmethod
-    def initialize(cls, scenario):
-        """Set up *scenario* with required sets and parameters for MESSAGE.
-
-        See Also
-        --------
-        :data:`MESSAGE_ITEMS`
-        """
-        # Initialize the ixmp items
-        cls.initialize_items(scenario, MESSAGE_ITEMS)
-
-    @staticmethod
-    def enforce(scenario):
-        """Enforce data consistency in `scenario`."""
-        # No-op; implemented in MESSAGE sub-class, below
-        # TODO make this an optional method of the ixmp.model.base.Model abstract class
 
     def __init__(self, name=None, **model_options):
         # Update the default options with any user-provided options
@@ -387,6 +377,44 @@ class GAMSModel(ixmp.model.gams.GAMSModel):
         return result
 
 
+def _check_structure(scenario):
+    """Check dimensionality of some items related to the storage representation.
+
+    Yields a sequence of 4-tuples:
+
+    1. Item name.
+    2. Item ix_type.
+    3. Number of data points in the item; -1 if it does not exist in `scenario`.
+    4. A warning/error message, *if* the index names/sets do not match those in
+       `MESSAGE_ITEMS` and the item contains data. Otherwise, the message is an empty
+       string.
+    """
+    if scenario.has_solution():
+        return
+
+    # NB could rename this e.g. _check_structure_0 if there are multiple such methods
+    for name in ("storage_initial", "storage_self_discharge", "map_tec_storage"):
+        info = MESSAGE_ITEMS[name]
+        message = ""
+
+        try:
+            # Retrieve the index names and data length of the item
+            idx_names = tuple(scenario.idx_names(name))
+            N = len(getattr(scenario, info["ix_type"])(name))
+        except KeyError:
+            N = -1  # Item does not exist
+        else:
+            # Item exists
+            expected_names = info.get("idx_names", info["idx_sets"])
+            if expected_names != idx_names and N > 0:
+                message = (
+                    f"{info['ix_type']} {name!r} has data with dimensions {idx_names!r}"
+                    f" != {expected_names!r} and cannot be solved; try expand_dims()"
+                )
+        finally:
+            yield name, info["ix_type"], N, message
+
+
 class MESSAGE(GAMSModel):
     """Model class for MESSAGE."""
 
@@ -395,6 +423,12 @@ class MESSAGE(GAMSModel):
     @staticmethod
     def enforce(scenario):
         """Enforce data consistency in `scenario`."""
+        # Raise an exception if any of the storage items have incorrect dimensions, i.e.
+        # non-empty error messages
+        messages = list(filter(None, [msg for *_, msg in _check_structure(scenario)]))
+        if messages:
+            raise ValueError("\n".join(messages))
+
         # Check masks ("mapping sets") that indicate which elements of corresponding
         # parameters are active/non-zero. Note that there are other masks currently
         # handled in JDBCBackend. For the moment, this code does not backstop that
@@ -415,6 +449,31 @@ class MESSAGE(GAMSModel):
             with scenario.transact(f"Enforce consistency of {set_name} and {par_name}"):
                 scenario.remove_set(set_name, existing)
                 scenario.add_set(set_name, expected)
+
+    @classmethod
+    def initialize(cls, scenario):
+        """Set up *scenario* with required sets and parameters for MESSAGE.
+
+        See Also
+        --------
+        :data:`MESSAGE_ITEMS`
+        """
+        # Check for storage items that may contain incompatible data or need to be
+        # re-initialized
+        state = None
+        for name, ix_type, N, message in _check_structure(scenario):
+            if len(message):
+                warn(message)  # Existing, incompatible data → conspicuous warning
+            elif N == 0:
+                # Existing, empty item → remove, even if it has the correct dimensions.
+                state = maybe_check_out(scenario, state)
+                getattr(scenario, f"remove_{ix_type}")(name)
+
+        # Initialize the ixmp items for MESSAGE
+        cls.initialize_items(scenario, MESSAGE_ITEMS)
+
+        # Commit if anything was removed
+        maybe_commit(scenario, state, f"{cls.__name__}.initialize")
 
 
 class MACRO(GAMSModel):
@@ -470,3 +529,8 @@ class MESSAGE_MACRO(MESSAGE, MACRO):
 
         # Append to the prepared solve_args
         self.solve_args.extend(mm_iter_args)
+
+    @classmethod
+    def initialize(cls, scenario, with_data=False):
+        MESSAGE.initialize(scenario)
+        MACRO.initialize(scenario, with_data)